Automatically reload audio when needed
In case we're not working with the preferred card, we check if the
preferred card is available when user does something (get/set the
mute/volume).

We refrain from doing that too often though, since it involves querying
alsa for the list of soundcard available, and then doing string
comparison on every element of this list. It wouldn't be wise to do that
for every user action: when the user is scrolling, that could be a bit
heavy.

So we maintain a timestamp and force an minimum interval between two
checks. At the moment I set it to 10 seconds.